chore: update pathname after renaming some file
